# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ProviEmplea API",
        "description": "API REST para la plataforma de empleo ProviEmplea. Permite gestionar talentos, empresas y procesos de selección con curriculum ciego.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Servidor de desarrollo local"
        }
    ],
    "paths": {
        "/admin/contactos": {
            "get": {
                "tags": [
                    "Administración"
                ],
                "summary": "Listar solicitudes de contacto",
                "description": "Retorna todas las solicitudes de contacto entre empresas y talentos. Límite: 60 requests/minuto.",
                "operationId": "getContactos",
                "parameters": [
                    {
                        "name": "estado",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pendiente",
                                "contactado",
                                "entrevista",
                                "seleccionado",
                                "no-seleccionado",
                                "proceso-cerrado"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContactoSolicitado"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Demasiadas solicitudes. Límite: 60 req/min. Espere el tiempo indicado en el header Retry-After."
                    }
                }
            },
            "post": {
                "tags": [
                    "Administración"
                ],
                "summary": "Registrar solicitud de contacto",
                "operationId": "crearContacto",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "empresa_id",
                                    "persona_id"
                                ],
                                "properties": {
                                    "empresa_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "persona_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "notas_admin": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contacto creado"
                    },
                    "409": {
                        "description": "Ya existe solicitud activa"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
        "/admin/contactos/{id}/estado": {
            "patch": {
                "tags": [
                    "Administración"
                ],
                "summary": "Actualizar estado del proceso",
                "operationId": "actualizarEstado",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "estado"
                                ],
                                "properties": {
                                    "estado": {
                                        "type": "string",
                                        "enum": [
                                            "pendiente",
                                            "contactado",
                                            "entrevista",
                                            "seleccionado",
                                            "no-seleccionado",
                                            "proceso-cerrado"
                                        ]
                                    },
                                    "notas_admin": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estado actualizado"
                    },
                    "404": {
                        "description": "Contacto no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ContactoSolicitado": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "empresa_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "persona_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "estado": {
                        "type": "string",
                        "example": "pendiente"
                    },
                    "notas_admin": {
                        "type": "string",
                        "nullable": true
                    },
                    "fecha_contacto": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "fecha_entrevista": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "fecha_resultado": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "rateLimitHeaders": {
                "type": "apiKey",
                "name": "X-RateLimit-Limit",
                "in": "header"
            }
        }
    },
...
}